fix: dont make sink async
The sink function was marked as async even though it took no async
action. As it also throws an error this can create problems
downstream if the caller is not catching the async error. While
consumers should account for the error, since this is not an async
function it unncessarily complicates error handling as intermediate
libraries that could do things synchronously and bubble the error up,
now need to handle everything async.